+++ tests/BaseTestCase.php
@@ -21,6 +21,14 @@ class BaseTestCase extends TestCase
             Route::get('/another', function () {
                 return 'Another value: ' . request('test');
             });
+            Route::get('/error', function () {
+                return response('Error: ' . request('test'), 500);
+            });
+            Route::get('/validation', function () {
+                request()->validate([
+                    'name' => 'required'
+                ]);
+            });
         });
     }